chore: split types and commented fields
