+++ lib/src/types/common.ts
@@ -1,0 +1,5 @@
+export type CurrencyCode = 'USD' | 'ARS' | 'BRL'
+
+export type PageResponse<T> = {
+  results: T[]
+}
