+++ lib/src/types/index.ts
@@ -1,0 +1,12 @@
+// Type definitions for Pluggy
+// Definitions by: Pluggy <https://github.com/pluggyai>
+
+export * from './common'
+export * from './connector'
+export * from './category'
+export * from './item'
+export * from './account'
+export * from './investment'
+export * from './transaction'
+export * from './identity'
+export * from './webhook'
